requirements.txt
Flask
python-docx
weasyprint
